requirements.txt
flask
dash
dash-bootstrap-components
dash-leaflet
geopandas
shapely
pandas
numpy
requests
census
python-dotenv
openai
pyshp
pymongo